Conflicts:
	gen/android/support/v7/appcompat/R.java
	gen/com/example/groupname/R.java

+++ src/com/example/groupname/SendDattActivity.java
@@ -259,7 +259,7 @@ public class SendDattActivity extends ActionBarActivity implements
 			point = settings.getInt("point", point);
 			point = point + 1;
 			SharedPreferences.Editor editor = settings.edit();
-			editor.putLong("point", point);
+			editor.putInt("point", point);
 			editor.commit();
 
 			Intent intent = new Intent(this, HaveAntidoteActivity.class);
